\name{bios666r-package}
\alias{bios666r-package}
\alias{bios666r}
\docType{package}
\title{
  R/Rcpp code for BIOSTAT666 class
}
\description{
  This repository contains a collection of R and Rcpp code for BIOSTAT666
  in the University of Michigan.
}
\author{
Hyun Min Kang

Maintainer: Hyun Min Kang <hmkang@umich.edu>
}
\keyword{ package }
