% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{emHaplotyping}
\alias{emHaplotyping}
\title{E-M haplotyping with a limited number of markers}
\usage{
emHaplotyping(genoMat, possibleHaps, max_iter = 100L, tol = 1e-06)
}
\arguments{
\item{genoMat}{- (n.samples) x (n.variants) matrix of 0,1,2-encoded genotypes}

\item{possibleHaps}{- (n.haplotypes) x (n.variants) matrix of 0,1-encoded possible haplotypes}

\item{max_iter}{- maximum number of E-M iterations}

\item{tol}{- tolerance of relative log-likelihood changes}
}
\value{
A list containing
   * hap_freqs : estimated frequencies of possible haplotypes
   * records : matrix keeping track of iteration, log-likelihood, and haplotype frequencies
}
\description{
E-M haplotyping with a limited number of markers
}
