% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ibsMatrix}
\alias{ibsMatrix}
\title{Calculate IBS matrix efficiently}
\usage{
ibsMatrix(genoMat)
}
\arguments{
\item{genoMat}{- (n.variants) x (n.sample) matrix of 0,1,2-encoded genotypes}
}
\value{
A matrix containing IBS matrix (symmetric)
}
\description{
Calculate IBS matrix efficiently
}
