% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kingRobust}
\alias{kingRobust}
\title{Estimate kinship coefficient in heterogeneous populations}
\usage{
kingRobust(genoMat)
}
\arguments{
\item{genoMat}{- (n.variants) x (n.sample) matrix of 0,1,2-encoded genotypes}
}
\value{
A matrix containing pairwise kinship estimates matrix (symmetric)
}
\description{
Estimate kinship coefficient in heterogeneous populations
}
